CAR_LICENSE_PLATE_RECOGNITION/car_license_plate_recognition/backend/requirements.txt
Flask
flask-cors
pymongo
PyJWT
requests
python-dotenv
Pillow
pytesseract
Werkzeug
Werkzeug
opencv-python
gunicorn